Create LMFDB_imaging.magma
Script to image the curves in LMFDB.

Note: Curve 1707 in the atpyical data (105625.a.528125.1) will cause issues. We run into an error when applying FromAnalyticJacobian in our attempts to find simple quadratic torsion. It doesn't give an error but will simply run without making any (obvious) progress and stall the computation. It's not clear how to fix this currently.